EVENT: Generate N*N grid of square div
I need to add an event to the RESET button.
The event needs to send the user a popup asking for the number of squares per side.
I need this number to generate new grid in the same total space as before.

Solution
* Create a 'click' event to prompt for user input
* Limit the user input number to 100 and don't accept values which are not numbers
* Use this number to generate new grid in the same total space

Note
Merge after solving the spacing issues caused by borders, margins,...